backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
python-multipart>=0.0.9
pydantic>=2.6
pydantic-settings>=2.2
python-docx>=1.1
PyMuPDF>=1.24
openai>=1.30
anthropic>=0.25
tiktoken>=0.7
rank-bm25>=0.2.2
cryptography>=42
aiosqlite>=0.20
sse-starlette>=2.1
httpx>=0.27
